Refactor: move viewholder to quickactions package
